+++ GRDB/Core/DatabaseValue.swift
@@ -53,18 +53,18 @@ public struct DatabaseValue : Hashable {
     /// The NULL DatabaseValue.
     public static let Null = DatabaseValue(storage: .Null)
     
-    /// Returns a DatabaseValue storing an Integer.
-    public init(int64: Swift.Int64) {
+    /// Returns a DatabaseValue storing an Int64.
+    public init(_ int64: Swift.Int64) {
         storage = .Int64(int64)
     }
     
-    /// Returns a DatabaseValue storing an Double.
-    public init(double: Swift.Double) {
+    /// Returns a DatabaseValue storing a Double.
+    public init(_ double: Swift.Double) {
         storage = .Double(double)
     }
     
     /// Returns a DatabaseValue storing a String.
-    public init(string: Swift.String) {
+    public init(_ string: Swift.String) {
         storage = .String(string)
     }
     
@@ -72,7 +72,7 @@ public struct DatabaseValue : Hashable {
     ///
     /// SQLite cant' store zero-length blobs: if data has zero length, the
     /// result is NULL.
-    public init(data: NSData) {
+    public init(_ data: NSData) {
         if data.length == 0 {
             // SQLite cant' store zero-length blobs.
             storage = .Null